补充 Wrongs 获取返回数据的方法; 完善 _hsConfig 方法

+++ hongs-core/src/main/java/app/hongs/action/anno/Supply.java
@@ -15,5 +15,5 @@ import java.lang.annotation.Target;
 public @interface Supply {
     String  conf() default "default";
     String  form() default "";
-    short   mode() default -1;
+    byte    mode() default -1;
 }
